requirements.txt
huggingface-hub==0.36.2
pandas==2.2.3
faiss-cpu==1.13.1
requests==2.33.1
sentence-transformers==4.1.0
scikit-learn==1.7.1
youtube-search-python==1.6.6
opencv-python==4.10.0.84
httpx==0.27.1
streamlit==1.49.0
deepface==0.0.99
tensorflow==2.14.0
keras==2.14.0
mtcnn==1.0.0
Pillow==11.3.0
numpy==1.26.4
retina-face==0.0.17
gdown==6.0.0
fire==0.7.1
lightecc==0.0.6
lightdsa==0.0.3
lightphe==0.0.24
PySocks==1.7.1